% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustTonality.R
\name{adjustTonality}
\alias{adjustTonality}
\title{Adjust tonality between major and minor}
\usage{
adjustTonality(midi, what = "major")
}
\arguments{
\item{midi}{midi notes to convert tonalities of}

\item{what}{which tonality to go from and switch to the opposite}
}
\value{
numeric
}
\description{
Adjust tonality between major and minor
}
