% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_current.R
\name{analyze_current}
\alias{analyze_current}
\title{Produce analyses of each counterpoint choice}
\usage{
analyze_current(i, cpt_targets, cpt, cf)
}
\arguments{
\item{i}{the index of the current note}

\item{cpt_targets}{the set of vectors of possible counterpoint targets}

\item{cpt}{the vector of counterpoint notes}

\item{cf}{the vector of cantus firmus notes}
}
\value{
data.frame
}
\description{
{Produce analyses of each counterpoint choice}
}
