% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_exercise.R
\name{analyze_exercise}
\alias{analyze_exercise}
\title{Analyze a counterpoint (above) exercise}
\usage{
analyze_exercise(cpt, cf, voice = "above")
}
\arguments{
\item{cpt}{the vector of counterpoint}

\item{cf}{the vector of the cantus firmus}

\item{voice}{above or below the cantus firmus}
}
\value{
data.frame
}
\description{
{Creates a data.frame describing the characteristics of the exercise.}
}
