% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_exercise_above.R
\name{analyze_exercise_above}
\alias{analyze_exercise_above}
\title{Analyze a counterpoint (above) exercise}
\usage{
analyze_exercise_above(cpt, cf)
}
\arguments{
\item{cpt}{the vector of counterpoint}

\item{cf}{the vector of the cantus firmus}
}
\value{
data.frame
}
\description{
{Creates a data.frame describing the characteristics of the exercise.}
}
