% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterpoint_above.R
\name{check_for_non_steps}
\alias{check_for_non_steps}
\title{Check for previous skips or leaps}
\usage{
check_for_non_steps(i, current_piece)
}
\arguments{
\item{i}{the current index}

\item{current_piece}{output from analyze_exercise applied to current piece}
}
\value{
boolean
}
\description{
Check for previous skips or leaps
}
