% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord_prog.R
\name{chord_prog}
\alias{chord_prog}
\title{Generate a chord progression given a sequence of chords and durations}
\usage{
chord_prog(
  chords,
  durations,
  harmonics = c(1, 0, 0.05),
  n = 4,
  bpm = 120,
  sampleRate = 48000
)
}
\arguments{
\item{chords}{list of chords, each element a MIDI vector representation of a chord}

\item{durations}{vector of durations}

\item{harmonics}{vector or (not implemented yet) list of harmonics for each chord}

\item{n}{number of beats in a measure}

\item{bpm}{beats per minute}

\item{sampleRate}{sample rate}
}
\value{
vector
}
\description{
Generate a chord progression given a sequence of chords and durations
}
