% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_harmonies.R
\name{classify_harmonies}
\alias{classify_harmonies}
\title{Classify harmonies as either perfect and imperfect}
\usage{
classify_harmonies(harmony, exclude_fourth = TRUE)
}
\arguments{
\item{harmony}{the harmony or vector of harmonies to classify}

\item{exclude_fourth}{boolean whether to exclude fourths or not}
}
\value{
vector of strings
}
\description{
{Classify harmonies in z12 as either "perfect" or "imperfect".}
}
