% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_motion.R
\name{classify_motion}
\alias{classify_motion}
\title{Classify motions as "contrary", "direct", or "oblique"}
\usage{
classify_motion(cpt_step, cf_step)
}
\arguments{
\item{cpt_step}{the counterpoint step intervals}

\item{cf_step}{the cantus firmus step intervals}
}
\value{
string
}
\description{
{Classify motions based off their step signs}
}
