% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_motion.R
\name{classify_motion1}
\alias{classify_motion1}
\title{Classify motions as "contrary", "direct", or "oblique"}
\usage{
classify_motion1(cpt_step, cf_step)
}
\arguments{
\item{cpt_step}{the current counterpoint step interval}

\item{cf_step}{the current cantus firmus step interval}
}
\value{
string
}
\description{
{Classify motions based off their step signs}
}
