% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterpoint_above.R
\name{counterpoint_above}
\alias{counterpoint_above}
\title{Compose cantus firmus above}
\usage{
counterpoint_above(cf, mode = "dorian", root = 60)
}
\arguments{
\item{cf}{cantus firmus}

\item{mode}{mode of the melody}

\item{root}{root note in midi 127}
}
\value{
list
}
\description{
Compose cantus firmus above
}
