% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{dcor}
\alias{dcor}
\title{Difference between correlation}
\usage{
dcor(synth, real)
}
\arguments{
\item{synth}{\code{[data.frame]} Synthetic data}

\item{real}{\code{[data.frame]} Real data}
}
\description{
Returns the Frobenius norm of the difference between the 
correlation matrices for numeric columns in synthetic
and real data.
}
\examples{
dcor(iris, iris) # 0
dcor(synthesize(cars), cars)

}
\seealso{
Other measures: 
\code{\link{dmean}()},
\code{\link{pmse}()},
\code{\link{qa}()}
}
\concept{measures}
