% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_minimalism.R
\name{entryExitProcess}
\alias{entryExitProcess}
\title{Simulate a sample path of an entry-exit process}
\usage{
entryExitProcess(tt, Q)
}
\arguments{
\item{tt}{time-length to simulate for (for musical applications, in measures)}

\item{Q}{the generator matrix of the Markov process}
}
\value{
data.frame
}
\description{
{Simulate a modified birth-death process, we call an entry-exit process, as our intention
is to model the arrival and exiting of voices in a musical piece. Based off a BD process, the number of voices
can go up or down, and jump to rests and back.}
}
