% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_minimalism.R
\name{entryExitQ}
\alias{entryExitQ}
\title{Create a Q-matrix for an Entry exit process}
\usage{
entryExitQ(num_voices, bd_rates)
}
\arguments{
\item{num_voices}{number of voices (positive integer) in piece}

\item{bd_rates}{the modified birth-death process rates, see details}
}
\value{
matrix
}
\description{
{Populate a \eqn{Q}-matrix, otherwise called infinitesimal generator matrix of a modified
birth and death process that models the entry-exits of voices in a musical piece.}
}
\details{
{The argument \code{bd_rates} must be a named \code{list} with elements
\itemize{
\item \code{rest} the rate of transitions to rests
\item \code{down} the rate of transitions down a voice
\item \code{up} the rate of transitions up a voice
\item \code{rest_any} the rate of transitions from a rest to any number of voices
\item \code{full_any} the rate of transitions from full-band to any number of voices}}
}
