% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_species.R
\name{first_species}
\alias{first_species}
\title{First species counterpoint}
\usage{
first_species(cantus_firmus, root, mode, voice = "above")
}
\arguments{
\item{cantus_firmus}{the given cantus firmus a vector in midi 0:127}

\item{root}{the root note of the cantus firmus mode}

\item{mode}{the mode in the key}

\item{voice}{compose a counterpoint "above" or "below"}
}
\value{
list
}
\description{
{First species counterpoint in the strict style.}
}
