% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstSpeciesAbove.R
\name{firstSpeciesAbove}
\alias{firstSpeciesAbove}
\title{Compose first species counterpoint above a given cantus firmus}
\usage{
firstSpeciesAbove(cantusFirmus, key)
}
\arguments{
\item{cantusFirmus}{cantus firmus melody in 0-12 to write a counterpoint above}

\item{key}{the key being composed in}
}
\value{
list
}
\description{
{Compose first species counterpoint above a given cantus firmus in mod 12.}
}
