% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqToMidi.R
\name{freqToMidi}
\alias{freqToMidi}
\title{Convert from frequencies to midi numbers}
\usage{
freqToMidi(freq)
}
\arguments{
\item{freq}{the frequency in hertz}
}
\value{
numeric
}
\description{
{Converts frequencies in hertz to midi numbers in 0-127}
}
