% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugueOutline.R
\name{fugueOutline}
\alias{fugueOutline}
\title{Markov chain outline model of a fugue}
\usage{
fugueOutline(
  N,
  m,
  P.voice,
  P.transform,
  P.key,
  P.answer,
  keys = c(0, 2, 4, 9),
  answers = c(0, 4, 5, 7)
)
}
\arguments{
\item{N}{number of statements of the subject}

\item{m}{number of voices in the fugue}

\item{P.voice}{the probability transition matrix of the voice-chain: which voice states the subject at what time}

\item{P.transform}{the probability transition matrix of the transformations of the subject}

\item{P.key}{the probability transition matrix of the subject-keys}

\item{P.answer}{the probability transition matrix of the answer-keys}

\item{keys}{the subject-keys state space}

\item{answers}{the answer-keys state-space}
}
\value{
data.frame containing time, voice, key, transform
}
\description{
{Model a fugue outline as a random walk of a subject through different voices, keys, and transformations via 3 Markov chains}
}
