% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstSpeciesNew.R
\name{gen_cantus_firmus}
\alias{gen_cantus_firmus}
\title{Generate a random cantus firmus given a root and mode}
\usage{
gen_cantus_firmus(n = 7, root = 60, mode = "ionian")
}
\arguments{
\item{n}{number of notes in cantus firmus}

\item{root}{root number of the cantus firmus (in aboslute midi numbers)}

\item{mode}{mode name}
}
\value{
vector
}
\description{
Generate a random cantus firmus given a root and mode
}
