% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cpt.R
\name{gen_cpt}
\alias{gen_cpt}
\title{Generate the next counterpoint note}
\usage{
gen_cpt(prev_note, cf, prev_harmony, prev_interval, root, mode, decision_rule)
}
\arguments{
\item{prev_note}{the previous counterpoint note}

\item{cf}{the current cantus firmus note}

\item{prev_harmony}{the previous harmony interval}

\item{prev_interval}{the previous melodic interval}

\item{root}{the root of the exercise}

\item{mode}{the mode of the exercise}

\item{decision_rule}{the decision rule to use}
}
\value{
numeric
}
\description{
{Generate the next counterpoint note}
}
