% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_cpt_choices.R
\name{gen_cpt_choices}
\alias{gen_cpt_choices}
\title{Generate possible counterpoint choices}
\usage{
gen_cpt_choices(prev_note, prev_harmony, cf, root, mode)
}
\arguments{
\item{prev_note}{previous counterpoint note}

\item{prev_harmony}{previous harmony interval}

\item{cf}{previous cantus firmus note}

\item{root}{the root of the exercise}

\item{mode}{the mode of the exercise}
}
\value{
vector
}
\description{
{Generate the intersection of possible notes based off harmony formed with the current
cantus firmus and the melodic interval formed between the previous counterpoint note.}
}
