% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstSpeciesNew.R
\name{gen_legal_melodic_sequence}
\alias{gen_legal_melodic_sequence}
\title{Generate the next note in a cantus firmus}
\usage{
gen_legal_melodic_sequence(prev_note, root, mode)
}
\arguments{
\item{prev_note}{the previous note in the cf}

\item{root}{the root of the cf (absolute not mod 12)}

\item{mode}{the mode name}
}
\value{
numeric
}
\description{
Generate the next note in a cantus firmus
}
