% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstSpeciesNew.R
\name{get_dissonances}
\alias{get_dissonances}
\title{Return all dissonances}
\usage{
get_dissonances(exclude_fourth = FALSE)
}
\arguments{
\item{exclude_fourth}{whether to exclude fourths as dissonant or not}
}
\value{
vector
}
\description{
Return all dissonances
}
