% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstSpeciesNew.R
\name{get_first_species_legal_harmonies}
\alias{get_first_species_legal_harmonies}
\title{Return legal harmonies in first species}
\usage{
get_first_species_legal_harmonies()
}
\value{
vector
}
\description{
Return legal harmonies in first species
}
