% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_species_legal_melodic_intervals.R
\name{get_first_species_legal_melodic_intervals}
\alias{get_first_species_legal_melodic_intervals}
\title{Return the legal melodic intervals (independent of key) for first species}
\usage{
get_first_species_legal_melodic_intervals()
}
\value{
vector
}
\description{
{Return the legal melodic intervals (independent of key) for first species}
}
