% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstSpeciesNew.R
\name{get_imperfect_consonances}
\alias{get_imperfect_consonances}
\title{Return imperfect consonances}
\usage{
get_imperfect_consonances()
}
\value{
vector
}
\description{
Return imperfect consonances
}
