% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstSpeciesNew.R
\name{get_leaps}
\alias{get_leaps}
\title{Return the legal melodic leaps (independent of key) for first species}
\usage{
get_leaps()
}
\value{
vector
}
\description{
Return the legal melodic leaps (independent of key) for first species
}
