% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode_12.R
\name{get_mode_12}
\alias{get_mode_12}
\title{Get the mod 12 representation of a mode}
\usage{
get_mode_12(mode)
}
\arguments{
\item{mode}{name mode}
}
\value{
vector
}
\description{
{Return vector of the mode in mod 12, given the string name}
}
