% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skips.R
\name{get_skips}
\alias{get_skips}
\title{Return the legal melodic skips (independent of key) for first species}
\usage{
get_skips()
}
\value{
vector
}
\description{
{Return the legal melodic skips (independent of key) for first species}
}
