% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstSpeciesNew.R
\name{get_steps}
\alias{get_steps}
\title{Return the legal melodic steps (independent of key) for first species}
\usage{
get_steps()
}
\value{
vector
}
\description{
Return the legal melodic steps (independent of key) for first species
}
