% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKey.R
\name{getKey}
\alias{getKey}
\title{Convenient function for getting keys}
\usage{
getKey(what = "major")
}
\arguments{
\item{what}{what key to return major or minor}
}
\value{
vector
}
\description{
Convenient function for getting keys
}
