% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_above.R
\name{initialize_above}
\alias{initialize_above}
\title{Initialize a counterpoint above the cantus firmus}
\usage{
initialize_above(cf, mode, root)
}
\arguments{
\item{cf}{the cantus firmus vector}

\item{mode}{the mode to compose in}

\item{root}{the root note in mod 127 the exercise is composed in}
}
\value{
vector
}
\description{
{Initialize a counterpoint above the cantus firmus}
}
