% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstSpeciesNew.R
\name{interval_spans}
\alias{interval_spans}
\title{Get the span of each interval among the mod 127 numbers of midi encoding}
\usage{
interval_spans()
}
\value{
list of vectors
}
\description{
Get the span of each interval among the mod 127 numbers of midi encoding
}
