% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majorToMinor.R
\name{majorToMinor}
\alias{majorToMinor}
\title{Adjust tonality of a given set of midi notes}
\usage{
majorToMinor(midi)
}
\arguments{
\item{midi}{notes to convert from major to minor}
}
\value{
vector
}
\description{
{Brings all major thirds, 6ths and 7ths to minors}
}
