% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_minimalism.R
\name{markov_minimalism}
\alias{markov_minimalism}
\title{Generate an approximate minimalist piece via Markov processes}
\usage{
markov_minimalism(
  num_voices,
  piece_length,
  bd_rates,
  rep_probs,
  harmonicsP,
  state_space_info,
  nb = 4,
  bpm = 120
)
}
\arguments{
\item{num_voices}{number of voices in the piece, a positive integer}

\item{piece_length}{the duration of the piece, in measures}

\item{bd_rates}{the transition rates, see \code{entryExitQ} documentation}

\item{rep_probs}{the vector of repeat probabilities for each voice}

\item{harmonicsP}{parameters for harmonics, see \code{random_harmonics} documentation}

\item{state_space_info}{complete state-space input, see \code{minimal_piece} documentation}

\item{nb}{the number of beats per measure}

\item{bpm}{the beats per minute}
}
\value{
list
}
\description{
{Generate a minimalist composition, approximately, via Markov processes simulations}
}
