% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureToSeconds.R
\name{measureToSeconds}
\alias{measureToSeconds}
\title{Convert time from measures to seconds}
\usage{
measureToSeconds(rhythm, n = 4, bpm = 120)
}
\arguments{
\item{rhythm}{the rhythm to convert to seconds, in a fraction of a measure, i.e. 1, 1/2, 1/4, 1/8, 1/16,}

\item{n}{number of beats in a measure, the lower number in a time signature N/n}

\item{bpm}{the beats per minute}
}
\value{
numeric
}
\description{
Convert time from measures to seconds
}
