% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_minimalism.R
\name{melodyTransP}
\alias{melodyTransP}
\title{Create a list of Probability transition matrices for the DTMC of repeated melodies}
\usage{
melodyTransP(num_voices, rep_probs, state_lengths)
}
\arguments{
\item{num_voices}{number of voices, a positive integer, in the musical piece}

\item{rep_probs}{the vector of probabilities representing chances to repeat current melody}

\item{state_lengths}{the size of the state space per voice, i.e. the number of allotted melodies}
}
\value{
list of matrices
}
\description{
{Populate the probability-transition matrices for the DTMCs representing which melody
repeats and randomly switches from. The diagonals are constant, and the non-diagonal entries have equal weight.}
}
\details{
{See the paper for more details, but essential staying in the same state has high chance, while jumping to
any other melody has a lower but equal chance, as any other distinct melody.}
}
