% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiPitchInvert.R
\name{midiPitchInvert}
\alias{midiPitchInvert}
\title{Invert pitches over an axis in midi units}
\usage{
midiPitchInvert(midi, axis = 60)
}
\arguments{
\item{midi}{midi pitches to invert}

\item{axis}{the axis to invert over}
}
\value{
numeric or vector
}
\description{
Invert pitches over an axis in midi units
}
