% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midiToFreq.R
\name{midiToFreq}
\alias{midiToFreq}
\title{Convert from midi numbers to frequencies in hertz}
\usage{
midiToFreq(midi)
}
\arguments{
\item{midi}{the midi number in 0-127}
}
\value{
numeric
}
\description{
{Formula for converting between frequencies and midi}
}
