% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_minimalism.R
\name{minimal_piece}
\alias{minimal_piece}
\title{Generate a minimalist piece via Markov processes}
\usage{
minimal_piece(
  sample_path,
  comp_param,
  arrangement,
  harmonicsP,
  state_space_info
)
}
\arguments{
\item{sample_path}{the sample path of the entry-exit process}

\item{comp_param}{the composition parameters, see details}

\item{arrangement}{the arrangement input-data, see details}

\item{harmonicsP}{the harmonics input data, see details}

\item{state_space_info}{the state-space input data, see details}
}
\value{
list
}
\description{
Generate a minimalist piece via Markov processes
}
\details{
{The argument \code{comp_param} must be a named list containing,
\itemize{
\item \code{num_voices} the number of voices in the piece, a positive integer
\item \code{nb} the number of beats per measure
\item \code{bpm} the beats per minute}

The argument \code{arrangement} must be a named list with elements
\itemize{
\item \code{voice_entries}, the list of vectors of indexes for voices played in a given section
\item \code{melody_Plist} the list of probability transition matrices for the DTMC of melodies per voice
\item \code{waiting_times} the duration of each section, in measures.}

The argument \code{state_space_info} must be a named list with elements
\itemize{
\item \code{state_lengths}, number of melodies per voice
\item \code{melody_states}, the melody-state space for each voice
\item \code{rhythm_states}, the rhythm-state space for each voice
\item \code{lengthOfMelodies}, the duration of each melody per voice
}}
}
