% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_minimalism.R
\name{minimal_voice}
\alias{minimal_voice}
\title{Synthesize the sounds of a single voice in a given section}
\usage{
minimal_voice(
  initial_voice,
  P,
  W,
  n,
  nb,
  bpm,
  harmonicParams,
  state_lengths,
  melody_states,
  rhythm_states
)
}
\arguments{
\item{initial_voice}{the index of the initial voice}

\item{P}{the probability transition matrix determining the melodic transitions}

\item{W}{the duration of the entire section}

\item{n}{the number of steps to generate the \eqn{DTMC(P)} sample-path (make it large)}

\item{nb}{the number of beats per measure}

\item{bpm}{the beats per minute}

\item{harmonicParams}{the parameters for the harmonics generation (see details)}

\item{state_lengths}{the size of the state-spaces for each voice i.e. the number of melodies per voice}

\item{melody_states}{the list of state-spaces containing the pitch-tones in midi numbering}

\item{rhythm_states}{the list of state-spaces containing the rhythms of each melody in fractions of measures}
}
\value{
list containing the wave vector and the time-error between input section length and generated
}
\description{
{Synthesizes a single voice over a given section with model input}
}
\details{
{The algorithm generates a large DTMC and then uses the sample path up until the section duration is reached,
this allows for state-spaces of melodies of varying lengths.

The argument \code{harmonicParams} must be a named list containing:
\itemize{
\item \code{lambdaRange} the range of mean number of harmonics
\item \code{betaRange} the range of the beta distribution parameters}
}
}
