% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minorToMajor.R
\name{minorToMajor}
\alias{minorToMajor}
\title{Adjust tonality of a given set of midi notes}
\usage{
minorToMajor(midi)
}
\arguments{
\item{midi}{notes to convert from minor to major}
}
\value{
vector
}
\description{
{Brings all major thirds, 6ths and 7ths to minors}
}
