% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_durations.R
\name{poisson_durations}
\alias{poisson_durations}
\title{Exponential waiting times}
\usage{
poisson_durations(t, lambda)
}
\arguments{
\item{t}{total duration to simulate for}

\item{lambda}{mean number of notes per unit}
}
\value{
vector
}
\description{
{Durations of notes given by waiting times of a Poisson process}
}
\details{
{We assume from the start of the zero-th arrival, a note is played so the waiting
time to the first arrival is the duration of the first note. }
}
