% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{qa}
\alias{qa}
\title{Quality assurance for synthesized data}
\usage{
qa(real, n = 10)
}
\arguments{
\item{real}{\code{[data.frame]} A data set to be synthesized.}

\item{n}{\code{[integer]} Number of repetitions}
}
\value{
A \code{data.frame} with \code{n} rows and each column a quality measure.
}
\description{
Repeatedly synthesize a dataset, record a set of quality
measures for each repetition.
}
\examples{

qa(iris)

}
\seealso{
Other measures: 
\code{\link{dcor}()},
\code{\link{dmean}()},
\code{\link{pmse}()}
}
\concept{measures}
