% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_minimalism.R
\name{random_harmonics}
\alias{random_harmonics}
\title{Generate a list of random harmonics}
\usage{
random_harmonics(synthInput, lambdaRange, betaRange)
}
\arguments{
\item{synthInput}{I have no idea. TODO: remember what this is}

\item{lambdaRange}{range of mean-number of harmonics per tone}

\item{betaRange}{range of beta parameters for amplitude distributions}
}
\value{
list
}
\description{
{Generate, for each tone, a random list of harmonics, where the number of harmonics is
Poisson distributed, with a randomly chosen mean rate, and the amplitudes of the harmonics
are IID Beta-distributed RVs, with randomly chosen shape/scale parameters.}
}
