% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad_counterpoint.R
\name{remove_bad_direct_motion}
\alias{remove_bad_direct_motion}
\title{Remove perfect harmonies approached by direct motion}
\usage{
remove_bad_direct_motion(cpt_targets, analyses)
}
\arguments{
\item{cpt_targets}{set of counterpoint targets}

\item{analyses}{analyses of them}
}
\value{
vector
}
\description{
{Remove perfect harmonies approached by direct motion}
}
