% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad_counterpoint.R
\name{remove_non_steps}
\alias{remove_non_steps}
\title{Remove non-steps}
\usage{
remove_non_steps(cpt_targets, analyses)
}
\arguments{
\item{cpt_targets}{set of counterpoint targets}

\item{analyses}{analyses of them}
}
\value{
vector
}
\description{
{Remove non steps}
}
