% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad_counterpoint.R
\name{remove_same_steps}
\alias{remove_same_steps}
\title{Remove same-steps}
\usage{
remove_same_steps(i, cpt_targets, analyses, current_piece)
}
\arguments{
\item{i}{current index of counterpoint we want to write}

\item{cpt_targets}{set of counterpoint targets}

\item{analyses}{analyses of them}

\item{current_piece}{the latest composed piece}
}
\value{
vector
}
\description{
{Remove steps in same direction as previous leap}
}
