% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_one.R
\name{sample_one}
\alias{sample_one}
\title{Sample one element from a vector}
\usage{
sample_one(from)
}
\arguments{
\item{from}{the vector to sample from}
}
\value{
numeric
}
\description{
{Wrapper to \code{sample} to handle singleton sets.}
}
