% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_minimalism.R
\name{secondsToMeasure}
\alias{secondsToMeasure}
\title{Convert from seconds to measures}
\usage{
secondsToMeasure(seconds, nb, bpm)
}
\arguments{
\item{seconds}{time value in seconds}

\item{nb}{number of beats per measure}

\item{bpm}{number of beats per minute}
}
\value{
numeric
}
\description{
{Convert frmo seconds to measures using the basic formula relating seconds, number of
beats per measure and the BPM}
}
