% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_wav_player.R
\name{set_wave_player}
\alias{set_wave_player}
\title{Set the wave player and return the sample rate}
\usage{
set_wave_player(computer = "windows")
}
\arguments{
\item{computer}{"windows" or "mac"}
}
\value{
numeric
}
\description{
{Return the sample rate and set player for each computer. A wrapper to \code{setWavPlayer}}
}
