% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasticFugue.R
\name{stochasticFugue}
\alias{stochasticFugue}
\title{Generate a stochastic fugue based on Markov chains}
\usage{
stochasticFugue(
  subject,
  rhythms,
  N,
  registers,
  transition_matrices,
  subjectKeys,
  answerKeys,
  keySig = "major"
)
}
\arguments{
\item{subject}{the fugue subject, a vector of midi numbers in 0-127}

\item{rhythms}{the rhythms, either constant or variable}

\item{N}{the number of statements to generate}

\item{registers}{the registers of each voice}

\item{transition_matrices}{a list of transition matrices, see details}

\item{subjectKeys}{the state space for keys of the subject}

\item{answerKeys}{the state space for the keys of the answers}

\item{keySig}{the key signature of the subject}
}
\value{
vector
}
\description{
{Three Markov chains are generated to model an outline of a fugue, i.e. the path of the subject through different voices,
keys, and transformations. Then counterpoint is filled in on each statement for the other voices. Random harmonics are used.}
}
\details{
{transition_matrices must be a list of four probability transition matrices:
\itemize{
\item P.voice
\item P.transform
\item P.key
\item P.answer}}
}
