% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_teams.R
\name{syn_teams}
\alias{syn_teams}
\title{Synthesize preferences and availability into teams}
\usage{
syn_teams(
  students,
  min_size = 3,
  max_size = 6,
  preference_weight = 10,
  team_penalty = 2,
  time_limit = 2
)
}
\arguments{
\item{students}{The data frame student preferences and availability are stored}

\item{min_size}{The minimum number of students per team}

\item{max_size}{The maximum number of students per team}

\item{preference_weight}{The weight of preferences}

\item{team_penalty}{Penalty for adding teams}

\item{time_limit}{Time limit in minutes}
}
\value{
A tibble dataframe of teams based on preferences and availability
}
\description{
This function uses integer programming from the \code{ompr} package. Each student
provides up to 4 peers who they would prefer working with and 1 peer they'd
like to avoid working with as well as their availability for 2 out-of-class meeting
times. These variables are stored in a data frame and \code{syn_teams} creates teams
that maximize preferences while avoiding non-preferences and ensuring shared
availability.
}
\examples{
library(readr)
library(synthesizer)

students <- read_csv("students.csv")
teams <- syn_teams(students)

}
