% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesizer_package.R
\docType{package}
\name{synthesizer-package}
\alias{synthesizer}
\alias{synthesizer-package}
\title{synthesizer: Fast, Robust, and High-Quality Synthetic Data Generation with a Tuneable Privacy-Utility Trade-Off}
\description{
Synthesize numeric, categorical, mixed and time series data. Data circumstances including mixed (or zero-inflated) distributions and missing data patterns are reproduced in the synthetic data. A single parameter allows balancing between high-quality synthetic data that represents correlations of the original data and lower quality but more privacy safe synthetic data without correlations. Tuning can be done per variable or for the whole dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markvanderloo/synthesizer}
}

}
\author{
\strong{Maintainer}: Mark van der Loo \email{mark.vanderloo@gmail.com} (\href{https://orcid.org/0000-0002-9807-4686}{ORCID})

}
\keyword{internal}
