% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone.R
\name{tone}
\alias{tone}
\title{Produce a synthesized sound}
\usage{
tone(midi, rhythm, harmonics = 1, n = 4, bpm = 120, sampleRate = 48000)
}
\arguments{
\item{midi}{the midi note number for the tone to produce}

\item{rhythm}{the duration of the note in fractions of a measure}

\item{harmonics}{the amplitude of the harmonics}

\item{n}{number of beats in a measure}

\item{bpm}{the beats per minute}

\item{sampleRate}{the sample rate of the device typically windows is 48000 and mac osX is 44100}
}
\value{
numeric vector
}
\description{
{Wrapper to \code{seewave::synth()} function to produce a pure sine-wave tone. Additional parameters available for different wave forms, soon to be implemented.}
}
\details{
{The rhythms are converted to seconds via the number of beats per measure and bpm. Midi notes in 0-127 are converted to frequencies in Hertz.
The harmonics may either be a vector with leading component unity followed by amplitudes in the interval \eqn{(0,1)} or simply the number one.}
}
