% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tones.R
\name{tones}
\alias{tones}
\title{Synthesize a melody or chord given a sequence of midi notes and durations in measures}
\usage{
tones(
  midi,
  rhythms,
  harmonics = c(1, 0.5, 0.25),
  type = "melody",
  n = 4,
  bpm = 120,
  sampleRate = 48000
)
}
\arguments{
\item{midi}{vector of midi notes in 0-127}

\item{rhythms}{vector of rhythms in fractions of a measure, can be a constant rhythm}

\item{harmonics}{vector of harmonics or list of vectors of harmonics (of possibly different lengths)}

\item{type}{melody or chord}

\item{n}{number of beats in a measure}

\item{bpm}{the beats per minute}

\item{sampleRate}{the sample rate of the device}
}
\value{
vector
}
\description{
{Synthesizes many sounds and strings them together for a melody or combine them for a chord}
}
\details{
{Harmonics may either be a single vector or a list of vectors of different lengths. The first entry must be unity and the rest must be in the unit interval.}
}
