% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_rules.R
\name{writing_rules}
\alias{writing_rules}
\title{Decision rules for counterpoint}
\usage{
writing_rules(cpt_targets, prev_note, cf, decision_rule = "naive")
}
\arguments{
\item{cpt_targets}{collection of possible counterpoint targets}

\item{prev_note}{the previous counterpoint note}

\item{cf}{the current cantus firmus note}

\item{decision_rule}{the decision rule to use}
}
\value{
numeric
}
\description{
Decision rules for counterpoint
}
