% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z12_to_intervals.R
\name{z12_to_intervals}
\alias{z12_to_intervals}
\title{Convert from Z12 to music theory pitches}
\usage{
z12_to_intervals(z)
}
\arguments{
\item{z}{numeric interval in Z12}
}
\value{
string
}
\description{
{Convenient function for converting between Z12 representation and
music theory pitch symbols.}
}
