% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode_127.R
\name{get_mode_127}
\alias{get_mode_127}
\title{Get the midi mod 127 representation of a mode}
\usage{
get_mode_127(mode)
}
\arguments{
\item{mode}{name mode}
}
\value{
vector
}
\description{
{Get the mod 127 midi representation of a mode given the mode name}
}
