% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_teams.R
\name{syn_teams}
\alias{syn_teams}
\title{Pull Data from a Database and dump it into a Tibble}
\usage{
syn_teams(
  students,
  min_size = 3,
  max_size = 6,
  preference_weight = 10,
  team_penalty = 2,
  time_limit = 2
)
}
\arguments{
\item{db}{The Database Connection created from \code{SQL()}}

\item{query}{The SQL query you want to run}
}
\value{
A Tibble Dataframe of the results from \code{query}
}
\description{
This is a wrapper function of \code{DBI::dbGetQuery()} meant to minic the
\code{db.get_dataframe()} function in gestalten. Instead of dumping the data to a
pandas dataframe, we're dumping it to a tibble, which is the R version of a
tidy dataframe. It has better printing properties, is lazy in its column
encoding (which is a good thing), and is the way we should view data in R.
}
\examples{
\dontrun{
library(gestaltenr)
app <- Gestalten(config_file = "config.local.yml")
db_ensigndw <- SQL(app = app, db_key = "ENSIGNDW")
data <- get_dataframe(db_ensigndw, "SELECT TOP 100 * FROM EnsignDW.dbo.facPCCPDPM")
}

}
