% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesizer_package.R
\docType{package}
\name{synthesizer-package}
\alias{synthesizer}
\alias{synthesizer-package}
\title{synthesizer: Synthesize Data Based on Empirical Quantile Functions and Rank Order Matching}
\description{
Data is synthesized using a combination of inverse transform sampling using the empirical quantile functions for each variable, and then copying the rank order structure from the original dataset. The syntesizer method has a tunable parameter allowing to gradually move from realistic and possibly unsafe synthetic data to decorrelated data of less utility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markvanderloo/synthesizer}
}

}
\author{
\strong{Maintainer}: Mark van der Loo \email{mark.vanderloo@gmail.com} (\href{https://orcid.org/0000-0002-9807-4686}{ORCID})

}
\keyword{internal}
