% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesizer_package.R
\docType{package}
\name{synthesizer-package}
\alias{synthesizer}
\alias{synthesizer-package}
\title{synthesizer: Synthesize Data Based on Emperical Multivariate Distributions and Rank Order Matching}
\description{
Data is synthesized using a combination of inverse transform sampling from the emperical quantile functions for each variable, and then copying the rank order structure from the original dataset. The package also includes a number of functions to measure the utility of synthesized datasets.
}
\keyword{internal}
